Move task util test code into TaskSettingsPanel __test__ dir
